+++ R/tables.R
@@ -2,6 +2,8 @@
 #' 
 #' @param projectId The BigQuery project ID
 #' @param datasetId A datasetId within projectId
+#' @param maxResults Number of results to return, default \code{1000}
+#' @param pageToken The tableID to start listing from, for more than 1000 result paging
 #' 
 #' @return dataframe of tables in dataset
 #' 
@@ -13,23 +15,34 @@
 #' 
 #' @family bigQuery meta functions
 #' @export
-bqr_list_tables <- function(projectId, datasetId){
+bqr_list_tables <- function(projectId, datasetId, maxResults = 1000, pageToken = ""){
   
   l <- googleAuthR::gar_api_generator("https://www.googleapis.com/bigquery/v2",
                                       "GET",
                                       path_args = list(projects = projectId,
                                                        datasets = datasetId,
                                                        tables = ""),
-                                      data_parse_function = function(x) {
-                                        d <- x$tables
-                                        out <- data.frame(id = d$id,
-                                                          projectId = d$tableReference$projectId,
-                                                          datasetId = d$tableReference$datasetId,
-                                                          tableId = d$tableReference$tableId)
-                                        
-                                      })
-  l(path_arguments = list(projects = projectId, 
-                          datasets = datasetId))
+                                      pars_args = list(maxResults = maxResults,
+                                                       pageToken = pageToken),
+                                      data_parse_function = parse_bqr_list_tables)
+  
+  out <- l(path_arguments = list(projects = projectId, 
+                                 datasets = datasetId))
+  
+  # if(!is.null(out$content$nextPageToken)){
+  #   npt <- out$content$nextPageToken
+  #   myMessage("Paging through results: ", npt, level = 2)
+  # }
+  
+  out
+}
+
+parse_bqr_list_tables <- function(x) {
+  d <- x$tables
+  data.frame(id = d$id,
+             projectId = d$tableReference$projectId,
+             datasetId = d$tableReference$datasetId,
+             tableId = d$tableReference$tableId, stringsAsFactors = FALSE)
   
 }
 
@@ -110,16 +123,26 @@ bqr_table_data <- function(projectId, datasetId, tableId,
 #' @param datasetId A datasetId within projectId.
 #' @param tableId Name of table you want.
 #' @param template_data A dataframe with the correct types of data
+#' @param timePartitioning Whether to create a partioned table
+#' @param expirationMs If a partioned table, whether to have an expiration time on the data. The default \code{0} is no expiration.
 #' 
 #' @return TRUE if created, FALSE if not.  
 #' 
 #' @details 
 #' 
-#' Creates a BigQuery table 
+#' Creates a BigQuery table.
+#' 
+#' If setting \code{timePartioning} to \code{TRUE} then the table will be a 
+#'   \href{partioned table}{https://cloud.google.com/bigquery/docs/creating-partitioned-tables}
 #' 
 #' @family bigQuery meta functions
 #' @export
-bqr_create_table <- function(projectId, datasetId, tableId, template_data){
+bqr_create_table <- function(projectId, 
+                             datasetId, 
+                             tableId, 
+                             template_data,
+                             timePartitioning = FALSE,
+                             expirationMs = 0L){
   
   l <- googleAuthR::gar_api_generator("https://www.googleapis.com/bigquery/v2",
                                       "POST",
@@ -127,6 +150,12 @@ bqr_create_table <- function(projectId, datasetId, tableId, template_data){
                                                        datasets = datasetId,
                                                        tables = "")
                                       )
+  expirationMs <- as.integer(expirationMs)
+  timeP <- NULL
+  if(timePartitioning){
+    if(expirationMs == 0) expirationMs <- NULL
+    timeP <- list(type = "DAY", expirationMs = expirationMs)
+  }
   
   config <- list(
         schema = list(
@@ -136,26 +165,29 @@ bqr_create_table <- function(projectId, datasetId, tableId, template_data){
           projectId = projectId,
           datasetId = datasetId,
           tableId = tableId
-        )
+        ),
+        timePartitioning = timeP
   )
   
- req <- try(l(path_arguments = list(projects = projectId, 
-                          datasets = datasetId),
-           the_body = config), silent = TRUE)
- 
- if(is.error(req)){
-   if(grepl("Already Exists", error.message(req))){
-     message("Table exists: ", tableId, "Returning FALSE")
-     out <- FALSE
-   } else {
-     stop(error.message(req))
-   }
- } else {
-   message("Table created: ", tableId)
-   out <- TRUE
- }
- 
- out
+  config <- rmNullObs(config)
+  
+  req <- try(l(path_arguments = list(projects = projectId, 
+                                     datasets = datasetId),
+               the_body = config), silent = TRUE)
+  
+  if(is.error(req)){
+    if(grepl("Already Exists", error.message(req))){
+      message("Table exists: ", tableId, "Returning FALSE")
+      out <- FALSE
+    } else {
+      stop(error.message(req))
+    }
+  } else {
+    message("Table created: ", tableId)
+    out <- TRUE
+  }
+  
+  out
   
 }